replaced parmest.py with the new API and moved example
